Manually implement the Debug trait for FftError
The trait that get's used to display the error message when an
`unwrap()` is called on an `Error` is the Debug traits implementation.
It's much easier to understand the error when the explanation is written
out rather than trying to figure out what the Error variant tuple
members represents.

This changes the Debug trait's behaviour to match that of the Display
trait. I'm unsure why the Error trait requires the implementing type to
be Display as well as Debug. My understanding is that it's a message
supposed to help the end-user rather than the developer. The same
message would be appropriate for both cases AFAICT.