Return a boxed slice instead of vector from make_*
The number of elements of a vector can be changed after returned while a
boxed slice cannot. This is safer and more clearly communicates the
intention of the function to the user while taking up less space on the
stack.

This change renames make_*_vec to make_*_buffer and returns a boxed
slice instead of a vector.