Relax requirement for scratch buffer length
rustfft does not require the scratch buffer length to be an exact match
but rather just anything bigger than the requirement. This allows a
single large buffer to be re-used for fft calculations rather than
allocating a buffer for each operation.

This commit relaxes realfft's requirement to match that of rustfft.